Adđ: add create product and import order

+++ src/main/java/com/example/coffee_shop_chain_management/entity/DetailImportOrder.java
@@ -27,7 +27,7 @@ public class DetailImportOrder {
     Material material;
 
     @Column(name = "quantity", nullable = false)
-    Integer quantity;
+    Double quantity;
 
     @Column(name = "description", nullable = false)
     Double price;
